+++ webapp/sass/profile.scss
@@ -217,29 +217,6 @@ $pane-width: 6px;
         display: flex;
         justify-content: center;
       }
-
-      .flamegraph-legend {
-        display: flex;
-        align-items: center;
-        font-size: 11px;
-        justify-content: center;
-        margin-bottom: 10px;
-      }
-
-      .flamegraph-legend-list {
-        display: flex;
-        justify-content: center;
-      }
-
-      .flamegraph-legend-item {
-        width: 35px;
-        text-align: center;
-
-        &:first-child,
-        &:last-child {
-          width: 40px;
-        }
-      }
     }
 
     &.vertical-orientation {
